# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management API",
        "description": "Documentació oficial de l'API d'Inventari - Projecte Final",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/devices": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get list of devices with filters",
                "operationId": "a75d9e7e123258c1e54a0c79e9e2caff",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "owner_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "provider_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "device_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of devices",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Device"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Create a new device",
                "operationId": "6274e6101f84d18fb00e2cda281342f4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Device"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Device created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/devices/{id}": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Show device details",
                "operationId": "71f4ba9897bb3a10d3fde22ecf6a5558",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Update an existing device",
                "operationId": "ce1342dc6f1253b4f17981b87f659867",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "physical_location": {
                                        "type": "string",
                                        "example": "Repair Center"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "repair"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Screen damaged"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Devices"
                ],
                "summary": "Delete a device",
                "operationId": "14e173161243728b613b47fafae197e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Device": {
                "title": "Device",
                "description": "Device model definition",
                "required": [
                    "device_type",
                    "brand",
                    "model",
                    "serial_number",
                    "physical_location",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "device_type": {
                        "type": "string",
                        "example": "laptop"
                    },
                    "brand": {
                        "type": "string",
                        "example": "Dell"
                    },
                    "model": {
                        "type": "string",
                        "example": "XPS 15"
                    },
                    "serial_number": {
                        "type": "string",
                        "example": "SN-123456789"
                    },
                    "inventory_number": {
                        "type": "string",
                        "example": "INV-001",
                        "nullable": true
                    },
                    "physical_location": {
                        "type": "string",
                        "example": "Office 101"
                    },
                    "status": {
                        "description": "in_stock, assigned, repair, retired",
                        "type": "string",
                        "example": "in_stock"
                    },
                    "purchase_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "warranty_end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-01-01"
                    },
                    "has_warranty": {
                        "type": "boolean",
                        "example": true
                    },
                    "owner_id": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "provider_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}